common code for getting blobs
